Another llvm "fix", in quotes, because it might be bogus.
If 'a' is a global with an address_space(256), then this fix
allows us to use "(void *)(long)&a" in prebuilt structs.

It looks bogus because "(void *)&a" still doesn't work, with
clang giving the error "error: initializer element is not a
compile-time constant".  Too bad, we can always generate the
first variant in pypy and be happy.