fix; but anyway seems to go nowhere
Merging ranges in small_overflow_obj_ranges_add was basically the only way this
branch could improve sync-time. However, even with the quadratic, dead-slow
version of merging, only 8% of objs passed to this function allowed a merge
to happen (and the current version reaches ~7%).
In this light, the branch is probably useless. It also suggests that smallmalloc
often allocates in isolated slots of already-used pages. Meaning, if the
allocation pattern of smallmalloc does not change, there is little chance of
syncing continuous memory ranges instead of individual objects.