test and fix for rare bug when triggering minor GC in hashtable_lookup
It was possible that allocating the entry would trigger a minor GC. Then, there
is an old hashtable with a young entry that may not get traced in the next
collection.
I chose to make fall back to a preexisting entry in that case, but another way
would be to make stm_hashtable_write_entry ensure that there is an stm_write()
on the hobj if 'entry' is young.