do not check for safepoint in the become_inev(DONT_WAIT) case
there is a deadlock if stm_next_to_finalize() calls become_inevitable while
keeping the g_finalizers.lock; and _commit_finalizers() just spins while trying
to acquire it. In that case, the second thread will not enter a safepoint ever,
so all other threads are waiting. This commit *should* fix it by really not
waiting in become_inevitable when DONT_SLEEP is passed as the option.